+++ db/migrate/20200203140136_add_funding_organisation_details_to_activities.rb
@@ -1,0 +1,9 @@
+class AddFundingOrganisationDetailsToActivities < ActiveRecord::Migration[6.0]
+  def change
+    change_table :activities do |t|
+      t.string :funding_organisation_name
+      t.string :funding_organisation_reference
+      t.string :funding_organisation_type
+    end
+  end
+end